agents/sre-multi-agent/fast_agentcore/patterns/sre-four-agent/requirements.txt
# AgentCore Runtime dependencies
bedrock-agentcore==1.0.6
langgraph-checkpoint-aws==1.0.1
PyJWT[crypto]>=2.10.1

# LangGraph and LangChain for orchestration
langgraph==1.0.5
langchain-aws==1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data validation and configuration
pydantic>=2.5.0
python-dotenv>=1.0.0

# Additional dependencies for SRE agents
httpx>=0.27.0
